+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.52.0 (2024-12-04)
 
 - ENH: Return analysis dependencies from REST API
+- ENH: Entry-point route
 - BUG: Propagate dependency errors
 
 ## 1.51.0 (2024-11-25)